# Conflicts:
#	src/main/resources/templates/theme.html

+++ src/main/java/spring/app/service/impl/QuestionServiceImpl.java
@@ -48,10 +48,22 @@ public class QuestionServiceImpl implements QuestionService {
         questionDao.update(question);
     }
 
+    /*
+     * Метод производит удаление вопроса из темы,
+     * а затем смещает позиции нижестоящих вопросов на одну вверх
+     * */
     @Transactional
     @Override
     public void deleteQuestionById(Long id) {
-        questionDao.deleteById(id);
+        Question question = questionDao.getById(id);
+        if (question != null) {
+            Long themeIdOfQuestion = question.getTheme().getId();
+            int questionPosition = question.getPosition();
+            int maxPosition = questionDao.getQuestionMaxPositionByThemeId(themeIdOfQuestion);
+
+            questionDao.deleteById(id);
+            questionDao.shiftQuestionsPosition(themeIdOfQuestion, questionPosition + 1, maxPosition, -1);
+        }
     }
 
     @Override
@@ -77,12 +89,6 @@ public class QuestionServiceImpl implements QuestionService {
     @Override
     public QuestionDto getQuestionDtoById(Long id) {
         return questionDao.getQuestionDtoById(id);
-    }
-
-    @Override
-    @Transactional
-    public void deleteByQuestionTheme(Long themeId, Long questionId) {
-        questionDao.deleteByQuestionTheme(themeId, questionId);
     }
 
     /*
@@ -132,7 +138,4 @@ public class QuestionServiceImpl implements QuestionService {
         questionDao.update(currentQuestion);
         return true;
     }
-
-
 }
-
